Add more code

+++ src/configuration.rs
@@ -3,7 +3,7 @@ use serde_derive::Deserialize;
 
 #[derive(Clone, Deserialize)]
 pub struct GMConfiguration {
-    pub fps: f32,
+    pub fps: u32,
     pub window_title: String,
     pub screen_width: u32,
     pub screen_height: u32,
@@ -12,7 +12,7 @@ pub struct GMConfiguration {
 impl GMConfiguration {
     pub fn new() -> Self {
         Self {
-            fps: 60.0,
+            fps: 60,
             window_title: "GreenMoon2D".to_string(),
             screen_width: 1024,
             screen_height: 768,
